Renaming AlertComponent to GlobalAlert for maintaining consistency
