+++ desktop/core/src/desktop/js/apps/jobBrowser/knockout/JobBrowserViewModel.js
@@ -18,7 +18,7 @@ import $ from 'jquery';
 import * as ko from 'knockout';
 
 import { getLastKnownConfig } from 'config/hueConfig';
-import { GLOBAL_ERROR_TOPIC } from 'reactComponents/AlertComponent/events';
+import { GLOBAL_ERROR_TOPIC } from 'reactComponents/GlobalAlert/events';
 import huePubSub from 'utils/huePubSub';
 import I18n from 'utils/i18n';
 import { withLocalStorage } from 'utils/storageUtils';
