+++ desktop/core/src/desktop/js/jquery/plugins/jquery.filechooser.js
@@ -18,7 +18,7 @@ import $ from 'jquery';
 import * as ko from 'knockout';
 import fileuploader from 'ext/fileuploader.custom';
 import qq from 'ext/fileuploader.custom.new.js';
-import { GLOBAL_ERROR_TOPIC, GLOBAL_INFO_TOPIC } from 'reactComponents/AlertComponent/events';
+import { GLOBAL_ERROR_TOPIC, GLOBAL_INFO_TOPIC } from 'reactComponents/GlobalAlert/events';
 import huePubSub from 'utils/huePubSub';
 import I18n from 'utils/i18n';
 import { hueLocalStorage } from 'utils/storageUtils';
