+++ desktop/core/src/desktop/js/ko/bindings/ko.dropzone.js
@@ -18,7 +18,7 @@ import Dropzone from 'dropzone';
 import $ from 'jquery';
 import * as ko from 'knockout';
 
-import { GLOBAL_ERROR_TOPIC, GLOBAL_INFO_TOPIC } from 'reactComponents/AlertComponent/events';
+import { GLOBAL_ERROR_TOPIC, GLOBAL_INFO_TOPIC } from 'reactComponents/GlobalAlert/events';
 import huePubSub from 'utils/huePubSub';
 import I18n from 'utils/i18n';
