+++ desktop/core/src/desktop/js/reactComponents/GlobalAlert/GlobalAlert.scss
