+++ desktop/core/src/desktop/js/reactComponents/GlobalAlert/events.ts
