+++ desktop/core/src/desktop/js/reactComponents/GlobalAlert/types.ts
