+++ desktop/core/src/desktop/js/reactComponents/imports.js
@@ -17,8 +17,8 @@ export async function loadComponent(name) {
     case 'AppBanner':
       return (await import('./AppBanner/AppBanner')).default;
 
-    case 'AlertComponent':
-      return (await import('./AlertComponent/AlertComponent')).default;
+    case 'GlobalAlert':
+      return (await import('./GlobalAlert/GlobalAlert')).default;
 
     case 'ReactExampleGlobal':
       return (await import('./ReactExampleGlobal/ReactExampleGlobal')).default;
